backend-project/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.5.0
PyPDF2==3.0.1
python-docx==0.8.11
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
chromadb==0.4.15
python-dotenv==1.0.0
waitress==3.0.0